pyproject.toml
[project]
name = "plants-disease-detection"
version = "0.2.0"
description = "Plants Disease Detection - Deep learning-based plant disease classification system"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    # Core PyTorch requirements
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "timm>=0.6.12",
    # Data processing
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.62.0",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "opencv-python>=4.5.0",
    "albumentations>=1.1.0",
    "scikit-image>=0.19.0",
    # Visualization and logging
    "tensorboard>=2.7.0",
    "tensorboardX>=2.5.0",
    # Utilities
    "psutil>=5.9.0",
    "PyYAML>=6.0",
    "requests>=2.27.0",
    "python-dotenv>=0.19.0",
    # Deep learning utilities
    "efficientnet-pytorch>=0.7.1",
    "lookahead==1.0",
    "ranger-adabelief==0.1.0",
    # File handling
    "rarfile>=4.0",
    "setuptools>=59.5.0",
    # Web scraping and data collection
    "selenium>=4.1.0",
    "webdriver-manager>=3.5.0",
    "beautifulsoup4>=4.10.0",
    "flickrapi>=2.4.0",
    # GUI (for data collection tools)
    "PyQt6>=6.4.0",
]

[project.optional-dependencies]
# Optional dependencies for experiment tracking and model export
extras = [
    "wandb>=0.12.0",  # for experiment tracking
    "onnx>=1.11.0",  # for model export
    "onnxruntime>=1.10.0",  # for model inference
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "dataset", "libs", "models", "utils"]

